# Replace rather than update Metal buffers