--- a/src/mbgl/mtl/render_pass.cpp
+++ b/src/mbgl/mtl/render_pass.cpp
@@ -48,7 +48,7 @@ RenderPass::RenderPass(CommandEncoder& commandEncoder_, const char* name, const 
     // Let the encoder pass along any groups pushed to it after this
     commandEncoder.trackRenderPass(this);
 
-    commandEncoder.context.clear();
+    commandEncoder.context.performCleanup();
 }
 
 RenderPass::~RenderPass() {