With #1831, small non-index buffers (<= 4KiB) are encoded directly into the render command encoder rather than referenced as buffers.  This should eliminate sequencing problems where the buffer is updated while the previous frame rendering is still using them.

In order to avoid the same problem with larger buffers (e.g., vertex attributes), this change makes buffer updates allocate a new Metal buffer instead of modifying an existing one.  Index buffers were already being replaced rather than updating.

The effect on performance in the benchmarks seems minimal, but cases where the vertex buffers are large and changing regularly (e.g., symbols with layout in dense areas) there may be a more significant effect.

The previous buffers will remain allocated until they are no longer referenced, so memory utilization will increase somewhat.

```
Before:
Average frame encoding time: 2.6382 ms
Average frame rendering time: 14.5671 ms

Average frame encoding time: 2.6830 ms
Average frame rendering time: 14.5455 ms

Average frame encoding time: 2.5971 ms
Average frame rendering time: 14.5503 ms

---

After:
Average frame encoding time: 2.6585 ms
Average frame rendering time: 14.5019 ms

Average frame encoding time: 2.6152 ms
Average frame rendering time: 14.5189 ms

Average frame encoding time: 2.6410 ms
Average frame rendering time: 14.5200 ms
```