@stefankarschti 2023-11-07
After some panning & zooming the street name labels are drawn in the wrong place. Maybe some UBO buffer is not getting updated and the matrix is wrong? 

https://github.com/maplibre/maplibre-native/assets/4198736/1c070374-e1e4-4948-a638-df5ce2c4dd7f