@TimSylvester 2023-11-07
About 1/3 of buffers above the 4k threshold in the benchmark, but only 11K of 1.8M = 0.6% of updates are to large buffers.

```
totalBuffers = 48322
totalBufferObjs = 15239
bufferUpdates = 1769897
bufferObjUpdates = 10878
bufferUpdateBytes = 296601088
uniformUpdateBytes = 107493168
```